correccion

+++ src/state.c
@@ -298,8 +298,8 @@ static void update_tail(game_state_t* state, unsigned int snum) {
 
   set_board_at(state, tail_row, tail_col, ' ');
 
-  int new_row = tail_row;
-  int new_col = tail_col;
+  int new_row = (int) tail_row;
+  int new_col = (int) tail_col;
 
   if (tail_char == 'w') new_row--;
   else if (tail_char == 's') new_row++;
@@ -333,9 +333,9 @@ void update_state(game_state_t* state, int (*add_food)(game_state_t* state)) {
     char next_tile = get_board_at(state, next_row, next_col);
 
     if (next_tile == '#' || is_snake(next_tile)) {
-      snake->live = false;
-      set_board_at(state, head_row, head_col, 'x'); 
-      continue;
+       snake->live = false;
+       set_board_at(state, head_row, head_col, 'x'); 
+       continue;
     }
 
     set_board_at(state, head_row, head_col, head_to_body(head_char));
@@ -344,11 +344,11 @@ void update_state(game_state_t* state, int (*add_food)(game_state_t* state)) {
     snake->head_col = next_col;
 
     if (next_tile == '*') {
-      if (add_food != NULL) {
-        add_food(state);
-      }
+       if (add_food != NULL) {
+           add_food(state); 
+       }
     } else {
-      update_tail(state, i);
+        update_tail(state, i); 
     }
   } 
 }
